+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## v0.0.3 - 2022-??-?? - Support the root
 
+* Internationalizing Domain Names in Applications (IDNA) support for zone and
+  record names (record values are not yet supported)
 * Enable SUPPORTS_ROOT_NS for management of root NS records. Requires
   octodns>=0.9.16.